refactor: housekeeping
